+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - fix some cases of the verb not removed from the input on execution (with a risk of accidental double execution)
 - add the `:filesystems` (short `:fs`) verb and state on windows (it was already present on linux and mac).
 - display files whose name isn't valid UTF-8 (they were previously ignored)
+- android executable is back
 
 ### v1.52.0 - 2025-11-01
 <a name="v1.52.0"></a>